+++ compat-patrouille-gradle-plugin/src/main/kotlin/compat/patrouille/CompatPatrouillePlugin.kt
@@ -5,10 +5,10 @@ import gratatouille.wiring.GPlugin
 import org.gradle.api.Project
 
 @GPlugin(id = "com.gradleup.compat.patrouille")
-fun compatPatrouillePlugin(target: Project) {
+fun tapmocPlugin(target: Project) {
   target.extensions.create(
     CompatPatrouilleExtension::class.java,
-    "compatPatrouille",
+    "tapmoc",
     CompatPatrouilleExtensionImpl::class.java,
     target,
   )