+++ CHANGELOG.md
@@ -8,7 +8,7 @@ _2025-08-20_
 
 A few bugfixes, upgrades and ergonomics improvements. Many thanks @OliverO2 and @Mr3zee for their feedback in this release.
 
-* [NEW] Add compatPatrouilleCheckRuntimeDependencies as a lifecycle task https://github.com/GradleUp/compat-patrouille/pull/23
+* [NEW] Add tapmocCheckRuntimeDependencies as a lifecycle task https://github.com/GradleUp/compat-patrouille/pull/23
 * [NEW] For JS and Wasm, add the KGP kotlin-stdlib instead of relying on `coreLibrariesVersion` https://github.com/GradleUp/compat-patrouille/pull/24
 * [FIX] Make checkApiDependencies lazier https://github.com/GradleUp/compat-patrouille/pull/20
 * [FIX] Fix KMP with multiple targets https://github.com/GradleUp/compat-patrouille/pull/21